Enable IPv6 dual protocol socket support on FreeBSD
It is possible to limit the jvm to default to IPv4 only sockets by
setting `java.net.preferIPv4Stack` to true. This is the default on
OpenBSD, but for otehr BSD's we follow the Java specification that
sockets should accept both IPv4 and IPv6 when the latter is available.

Sponsored by: The FreeBSD Foundation