Make os::current_stack_base_and_size arch independent
Apart from a workaround for some OSX versions on x86_64 architectures,
the implementation is identical for all architectures. This patch moves
it to the base os_bsd.cpp file so it's included regardless of target
architecture.

This should fix the build issues for the PPC FreeBSD port for the last
few OpenJDK versions.

Sponsored by: The FreeBSD Foundation